maxdepthを追加とファイルの存在確認

+++ setup.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-for file in `find $HOME/dotfiles -name '.*' -d 1 | grep -v 'dotfiles/.git$' | perl -nle 'm!dotfiles/(.+)$! and print $1'`; do
-    ln -s $HOME/dotfiles/$file $HOME/$file
+for file in `find $HOME/dotfiles -maxdepth 1 -name '.*' | grep -v 'dotfiles/.git$' | perl -nle 'm!dotfiles/(.+)$! and print $1'`; do
+    if [ -e $HOME/$file ]; then
+        echo "exists $HOME/$file"
+    else
+        ln -s $HOME/dotfiles/$file $HOME/$file
+    fi
 done
